Allow for preloading the suffixes file
Currently, every call to `Rippersnapper::parse` results in the suffix
file being loaded. It's a large(ish) file and causes a fair amount of
object allocation for a file that typically doesn't change often for an
application.

This updates the top-level `Rippersnapper` module to have three new
methods:

* `::load_suffixes` - creates a new instance of
  `Rippersnapper::SuffixFileReader` with an optional file path and
  assigns it to a class instance variable.
* `::unload_suffixes` - resets the class instance variable to `nil` to
  preserve the current behavior of loading the file each time.
* `::suffix_file_reader` - simple accessor to get to the class ivar.

Internally, `Rippersnapper::DomainParser` checks for the existence of
`Rippersnapper::suffix_file_reader` and uses that. If the module-level
reader isn't found it defaults to the original behavior of loading a new
instance and reading the file.

Corresponding specs have been updated and make liberal use of the
`::unload_suffixes` method to make sure existing tests don't use the
preloaded file.